Support GHC 9.6 in CI (#77)
* Support GHC 9.6 in CI

* Malke sure we import the correct stuff from mtl

* Simplify CI, don't run for ghc9.2

* Install hspec-discover earlier in CI

* Downgrade to souffle 2.2

* Downgrade to souffle 2.2

* Add dockerfile for building souffle 2.2

* Update bench.cpp for souffle 2.2

---------

+++ cbits/souffle/datastructure/BTree.h
@@ -624,7 +624,7 @@ protected:
                         }
                     }
 
-                    pos = (i > static_cast<unsigned>(other->numElements)) ? 0 : static_cast<unsigned>(i);
+                    pos = (i > other->numElements) ? 0 : i;
                     other->insert_inner(root, root_lock, pos, predecessor, key, newNode, locked_nodes);
 #else
                     other->insert_inner(root, root_lock, pos, predecessor, key, newNode);
@@ -1340,8 +1340,7 @@ public:
 
                 // split this node
                 auto old_root = root;
-                idx -= cur->rebalance_or_split(
-                        const_cast<node**>(&root), root_lock, static_cast<int>(idx), parents);
+                idx -= cur->rebalance_or_split(const_cast<node**>(&root), root_lock, idx, parents);
 
                 // release parent lock
                 for (auto it = parents.rbegin(); it != parents.rend(); ++it) {
@@ -1373,7 +1372,7 @@ public:
             assert(cur->numElements < node::maxKeys && "Split required!");
 
             // move keys
-            for (int j = static_cast<int>(cur->numElements); j > static_cast<int>(idx); --j) {
+            for (int j = cur->numElements; j > idx; --j) {
                 cur->keys[j] = cur->keys[j - 1];
             }
 
@@ -1961,7 +1960,7 @@ private:
         const int N = node::maxKeys;
 
         // divide range in N+1 sub-ranges
-        int64_t length = (b - a) + 1;
+        int length = (b - a) + 1;
 
         // terminal case: length is less then maxKeys
         if (length <= N) {
@@ -1978,7 +1977,7 @@ private:
 
         // recursive case - compute step size
         int numKeys = N;
-        int64_t step = ((length - numKeys) / (numKeys + 1));
+        int step = ((length - numKeys) / (numKeys + 1));
 
         while (numKeys > 1 && (step < N / 2)) {
             numKeys--;
